pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "msgraph-sdk"
# The SDK version
# x-release-please-start-version
version = "1.9.0"
# x-release-please-end
authors = [{name = "Microsoft", email = "graphtooling+python@microsoft.com"}]
description = "The Microsoft Graph Python SDK"
dependencies = [
    "azure-identity >=1.12.0",
    "microsoft-kiota-abstractions >=1.3.0,<2.0.0",
    "microsoft-kiota-authentication-azure >=1.0.0,<2.0.0",
    "microsoft-kiota-serialization-json >=1.3.0,<2.0.0",
    "microsoft-kiota-serialization-text >=1.0.0,<2.0.0",
    "microsoft-kiota-serialization-form >=0.1.0",
    "microsoft-kiota-serialization-multipart >=0.1.0",
    "microsoft-kiota-http >=1.0.0,<2.0.0",
    "msgraph_core >=1.0.0"

]
requires-python = ">=3.8"
license = {file = "LICENSE"}
readme = "README.md"
keywords = ["msgraph", "openAPI", "Microsoft", "Graph"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]

[project.optional-dependencies]
dev = ["yapf", "bumpver", "isort", "pylint", "pytest", "mypy"]

[project.urls]
homepage = "https://github.com/microsoftgraph/msgraph-sdk-python#readme"
repository = "https://github.com/microsoftgraph/msgraph-sdk-python"
documentation = "https://github.com/microsoftgraph/msgraph-sdk-python/docs"

[tool.flit.module]
name = "msgraph"

[tool.mypy]
warn_unused_configs = true
files = "msgraph"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "msgraph.generated"
ignore_errors = true

[tool.yapf]
based_on_style = "pep8"
dedent_closing_brackets = true
each_dict_entry_on_separate_line = true
column_limit = 100

[tool.isort]
profile = "hug"

[tool.pytest.ini_options]
pythonpath = [
  "msgraph"
]

[tool.bumpver]
current_version = "1.5.0"
version_pattern = "MAJOR.MINOR.PATCH[PYTAGNUM]"
commit_message = "bump version {old_version} -> {new_version}"
commit = true
tag = false
push = false

[tool.bumpver.file_patterns]
"pyproject.toml" = ['current_version = "{version}"', 'version = "{version}"']
"msgraph/_version.py" = ["{version}"]

[tool.poetry.packages]
include = ["msgraph"]